+++ main_program/Library/movie_booking_framework/valid_checker.py
@@ -80,29 +80,31 @@ def movie_seats_pointer_valid_check(movie_seats : list,x_pointer : int,y_pointer
         raise e
 
 
-def date_valid_check(date : str) -> bool:
+def date_valid_check(date : str) -> tuple[bool, str]:
+    date_format : str = "YYYY/MM/DD"
     def leap_year_check(year : int) -> bool:
         if year % 4 == 0 and year % 100 != 0: return True
         if year % 400 == 0: return True
         return False
 
-    if not re.fullmatch(r"(\d{4})/(\d{2})/(\d{2})", date): return False
+    if not re.fullmatch(r"(\d{4})/(\d{2})/(\d{2})", date): return False,date_format
 
     date_after_split = re.search(r"(\d{4})/(\d{2})/(\d{2})", date)
     year, month, day = date_after_split.groups()
     month_30_day : list =[4,6,9,11]
-    if int(month) > 12: return False
-    if int(day) > 31: return False
-    if int(month) in month_30_day and int(day) > 30:return False
-    if int(month) == 2 and not leap_year_check(int(year)) and int(day) > 28: return False
-    if int(month) == 2 and leap_year_check(int(year)) and int(day) > 29: return False
-    return True
+    if int(month) > 12: return False,date_format
+    if int(day) > 31: return False,date_format
+    if int(month) in month_30_day and int(day) > 30:return False,date_format
+    if int(month) == 2 and not leap_year_check(int(year)) and int(day) > 28: return False,date_format
+    if int(month) == 2 and leap_year_check(int(year)) and int(day) > 29: return False,date_format
+    return True,""
 
-def time_valid_check(time : str) -> bool:
-    if not re.fullmatch(r"(\d{2}):(\d{2})",time) : return False
+def time_valid_check(time : str) -> tuple[bool, str]:
+    time_format : str = "HH:MM"
+    if not re.fullmatch(r"(\d{2}):(\d{2})",time) : return False,time_format
 
     time_after_split = re.search(r"(\d{2}):(\d{2})",time)
     hour,minute = time_after_split.groups()
-    if int(hour) > 23: return False
-    if int(minute) > 59: return False
-    return True
+    if int(hour) > 23: return False,time_format
+    if int(minute) > 59: return False,time_format
+    return True,""